requirements.txt
pandas~=2.1.2
mlflow~=2.8.0
notebook==7.0.6
numpy~=1.26.1
scikit-learn~=1.3.2
matplotlib~=3.7.2
python-box==7.1.1
pyYAML~=6.0.1
tqdm
ensure~=1.0.3
joblib~=1.3.2
types-PyYAML
Flask~=3.0.0
Flask-Cors
pymongo~=4.5.0
python-dotenv~=1.0.0
setuptools~=68.2.2
plotly~=5.18.0
statsmodels~=0.14.0
python-dateutil~=2.8.2
streamlit~=1.28.2
requests~=2.31.0
seaborn~=0.12.2
-e .
